Rename folder to 'spell'
