+++ client/app/src/main/java/enchantedtowers/client/components/spell/Spell.java
@@ -1,4 +1,4 @@
-package enchantedtowers.client.components.enchantment;
+package enchantedtowers.client.components.spell;
 
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
